refactor(app): split DesignTokens.swift into per-namespace token files (#3007)
384-line file held four independent token namespaces; extract three:

- SpacingTokens.swift — RBSpacing, RBRadius (8-pt grid + radii)
- MetricsTokens.swift — RBMetrics scalar constants
- TypographyTokens.swift — RBFont shared fonts

DesignTokens.swift keeps the adaptive color helpers, semantic color
tokens, and the RBStatus.color extension. No behaviour change.
