+++ Sources/RunBot/DesignSystem/DesignTokens.swift
@@ -323,62 +323,3 @@ extension RBStatus {
     }
 }
 
-// MARK: - Spacing & Geometry Tokens
-
-/// Fixed spacing constants derived from an 8-pt grid. Use these instead of raw `CGFloat` literals.
-enum RBSpacing {
-    /// 2 pt — hairline gap between tightly packed elements.
-    static let xxs: CGFloat = 2
-    /// 4 pt — compact inner padding (e.g. badge insets).
-    static let xs: CGFloat = 4
-    /// 6 pt — tight gap between related elements.
-    static let sm: CGFloat = 6
-    /// 10 pt — default row horizontal padding.
-    static let md: CGFloat = 10
-}
-
-/// Corner-radius constants for consistent rounding across components.
-enum RBRadius {
-    /// 10 pt — standard card corner radius.
-    static let card: CGFloat = 10
-    /// 6 pt — small card or row corner radius.
-    static let small: CGFloat = 6
-}
-
-// MARK: - Metrics Tokens
-
-/// Miscellaneous scalar constants that don't belong to spacing, radius, or typography.
-enum RBMetrics {
-    /// Scale factor applied to `ProgressView` in the update action row.
-    /// 0.7 keeps the spinner visually consistent with the `.small` control size of adjacent buttons.
-    static let updateProgressScale: CGFloat = 0.7
-    /// Maximum width for the commit/PR title label in `ActionRowView` (#2130).
-    /// Caps the greedy `.layoutPriority(1)` text so it can't consume the full row width.
-    /// The existing `.lineLimit(1)` + `.truncationMode(.tail)` produce an ellipsis at this cap.
-    static let actionRowTitleMaxWidth: CGFloat = 160
-    /// Maximum width for the head-branch label in `ActionRowView`.
-    /// Kept narrower than `actionRowTitleMaxWidth` since branch names are lower priority (#1194).
-    static let actionRowBranchMaxWidth: CGFloat = 80
-}
-
-// MARK: - Typography Tokens
-
-/// Shared font constants. Prefer these over inline `.system(size:weight:design:)` calls.
-enum RBFont {
-    /// Caption-sized monospaced font — general-purpose code/metric labels.
-    static let mono: Font = .system(.caption, design: .monospaced)
-    /// 11 pt regular monospaced — small metric values.
-    static let monoSmall: Font = .system(size: 11, weight: .regular, design: .monospaced)
-    /// 13 pt medium — standard row/list label.
-    static let label: Font = .system(size: 13, weight: .medium)
-    /// 12.5 pt regular — section key labels.
-    static let sectionKey: Font = .system(size: 12.5, weight: .regular)
-    /// Alias for `sectionKey` — section header labels.
-    static let sectionHeader: Font = sectionKey
-    /// 9 pt semibold — uppercase section caption badges.
-    static let sectionCaption: Font = .system(size: 9, weight: .semibold)
-    /// 9 pt semibold monospaced — stat label (CPU, MEM, etc.).
-    static let statLabel: Font = .system(size: 9, weight: .semibold, design: .monospaced)
-    /// 10 pt regular monospaced — numeric stat value.
-    static let statValue: Font = .system(size: 10, weight: .regular, design: .monospaced)
-}
